updates to simulator

+++ Robust/TransSim/TEvent.java
@@ -1,24 +1,24 @@
 public class TEvent {
-  public int type;
+  public byte type;
   public int index;
   public int oid;
   public long time;
 
-  public TEvent(int type, long time) {
+  public TEvent(byte type, long time) {
     this.time=time;
     this.type=type;
     this.oid=-1;
     this.index=-1;
   }
 
-  public TEvent(int type, long time, int oid) {
+  public TEvent(byte type, long time, int oid) {
     this.time=time;
     this.type=type;
     this.oid=oid;
     this.index=-1;
   }
 
-  public TEvent(int type, long time, int oid, int index) {
+  public TEvent(byte type, long time, int oid, int index) {
     this.time=time;
     this.type=type;
     this.oid=oid;
